#16171: Preload kernels before receiving go message
Add a flag that lets brisc.cc start loading kernels before receiving a go
message. Fast dispatch ensures that that the flag will be set only after all
necessary program data is sent to the core.

This allows preparation for the following kernel (including loading NCRISC
IRAM, setting up CBs, and initializing local memory) to happen in parallel with
the round-trip to the dispatcher_s to sync up with the other kernels and ensure
that they're all launched at the same time.